Move kernelCommandListener into notebookCommandListener
