+++ src/notebooks/notebookCommandListener.ts
@@ -5,22 +5,50 @@ import '../platform/common/extensions';
 
 import { inject, injectable } from 'inversify';
 
-import { NotebookCellData, NotebookCellKind, NotebookEdit, NotebookRange } from 'vscode';
-import { IVSCodeNotebook, ICommandManager } from '../platform/common/application/types';
-import { IDataScienceCommandListener, IDisposableRegistry } from '../platform/common/types';
-import { Commands } from '../webviews/webview-side/common/constants';
+import {
+    ConfigurationTarget,
+    NotebookCellData,
+    NotebookCellKind,
+    NotebookEdit,
+    NotebookRange,
+    Uri,
+    workspace
+} from 'vscode';
+import { IVSCodeNotebook, ICommandManager, IApplicationShell } from '../platform/common/application/types';
+import { IConfigurationService, IDataScienceCommandListener, IDisposableRegistry } from '../platform/common/types';
+import { Commands, Telemetry } from '../webviews/webview-side/common/constants';
 import { chainWithPendingUpdates } from './execution/notebookUpdater';
 import { getNotebookMetadata } from './helpers';
 import { noop } from '../platform/common/utils/misc';
 import { NotebookCellLanguageService } from '../intellisense/cellLanguageService';
+import { DisplayOptions } from '../kernels/displayOptions';
+import { KernelConnector } from '../kernels/kernelConnector';
+import { IKernel, IKernelProvider } from '../kernels/types';
+import { getDisplayPath } from '../platform/common/platform/fs-paths';
+import { DataScience } from '../platform/common/utils/localize';
+import { endCellAndDisplayErrorsInCell } from '../platform/errors/errorUtils';
+import { traceInfoIfCI, traceInfo } from '../platform/logging';
+import { sendTelemetryEvent } from '../telemetry';
+import { trackKernelResourceInformation } from '../telemetry/telemetry';
+import { getAssociatedNotebookDocument } from './controllers/kernelSelector';
+import { INotebookControllerManager } from './types';
+import { IDataScienceErrorHandler } from '../platform/errors/types';
+import { IServiceContainer } from '../platform/ioc/types';
 
 @injectable()
 export class NotebookCommandListener implements IDataScienceCommandListener {
+    private kernelInterruptedDontAskToRestart: boolean = false;
     constructor(
         @inject(IDisposableRegistry) private disposableRegistry: IDisposableRegistry,
         @inject(IVSCodeNotebook) private notebooks: IVSCodeNotebook,
         @inject(ICommandManager) private commandManager: ICommandManager,
-        @inject(NotebookCellLanguageService) private readonly languageService: NotebookCellLanguageService
+        @inject(NotebookCellLanguageService) private readonly languageService: NotebookCellLanguageService,
+        @inject(IApplicationShell) private applicationShell: IApplicationShell,
+        @inject(IConfigurationService) private configurationService: IConfigurationService,
+        @inject(IKernelProvider) private kernelProvider: IKernelProvider,
+        @inject(INotebookControllerManager) private notebookControllerManager: INotebookControllerManager,
+        @inject(IDataScienceErrorHandler) private errorHandler: IDataScienceErrorHandler,
+        @inject(IServiceContainer) private serviceContainer: IServiceContainer
     ) {}
 
     public register(commandManager: ICommandManager): void {
@@ -44,6 +72,44 @@ export class NotebookCommandListener implements IDataScienceCommandListener {
         );
         this.disposableRegistry.push(
             this.commandManager.registerCommand(Commands.NotebookEditorExpandAllCells, this.expandAll, this)
+        );
+        this.disposableRegistry.push(
+            commandManager.registerCommand(
+                Commands.NotebookEditorInterruptKernel,
+                (context?: { notebookEditor: { notebookUri: Uri } } | Uri) => {
+                    if (context && 'notebookEditor' in context) {
+                        void this.interruptKernel(context?.notebookEditor?.notebookUri);
+                    } else {
+                        void this.interruptKernel(context);
+                    }
+                }
+            )
+        );
+        this.disposableRegistry.push(
+            commandManager.registerCommand(
+                Commands.NotebookEditorRestartKernel,
+                (context?: { notebookEditor: { notebookUri: Uri } } | Uri) => {
+                    if (context && 'notebookEditor' in context) {
+                        void this.restartKernel(context?.notebookEditor?.notebookUri);
+                    } else {
+                        void this.restartKernel(context);
+                    }
+                }
+            )
+        );
+        this.disposableRegistry.push(
+            commandManager.registerCommand(
+                Commands.InterruptKernel,
+                (context?: { notebookEditor: { notebookUri: Uri } }) =>
+                    this.interruptKernel(context?.notebookEditor?.notebookUri)
+            )
+        );
+        this.disposableRegistry.push(
+            commandManager.registerCommand(
+                Commands.RestartKernel,
+                (context?: { notebookEditor: { notebookUri: Uri } }) =>
+                    this.restartKernel(context?.notebookEditor?.notebookUri)
+            )
         );
     }
 
@@ -110,5 +176,140 @@ export class NotebookCommandListener implements IDataScienceCommandListener {
                 edit.set(document.uri, [NotebookEdit.updateCellMetadata(index, metadata)]);
             });
         }).then(noop, noop);
+    }
+
+    public async interruptKernel(notebookUri: Uri | undefined): Promise<void> {
+        const uri =
+            notebookUri ??
+            window.activeNotebookEditor?.notebook.uri ??
+            this.interactiveWindowProvider?.activeWindow?.notebookUri ??
+            (window.activeTextEditor?.document.uri &&
+                this.interactiveWindowProvider?.get(window.activeTextEditor.document.uri)?.notebookUri);
+        const document = workspace.notebookDocuments.find((document) => document.uri.toString() === uri?.toString());
+
+        if (document === undefined) {
+            return;
+        }
+        traceInfoIfCI(`Interrupt kernel command handler for ${getDisplayPath(document.uri)}`);
+
+        const kernel = this.kernelProvider.get(document.uri);
+        if (!kernel) {
+            traceInfo(`Interrupt requested & no kernel.`);
+            return;
+        }
+        await this.wrapKernelMethod('interrupt', kernel);
+    }
+
+    private async restartKernel(notebookUri: Uri | undefined) {
+        const uri =
+            notebookUri ??
+            window.activeNotebookEditor?.notebook.uri ??
+            this.interactiveWindowProvider?.activeWindow?.notebookUri ??
+            (window.activeTextEditor?.document.uri &&
+                this.interactiveWindowProvider?.get(window.activeTextEditor.document.uri)?.notebookUri);
+        const document = workspace.notebookDocuments.find((document) => document.uri.toString() === uri?.toString());
+
+        if (document === undefined) {
+            return;
+        }
+
+        sendTelemetryEvent(Telemetry.RestartKernelCommand);
+        const kernel = this.kernelProvider.get(document.uri);
+
+        if (kernel) {
+            trackKernelResourceInformation(kernel.resourceUri, { restartKernel: true });
+            if (await this.shouldAskForRestart(document.uri)) {
+                // Ask the user if they want us to restart or not.
+                const message = DataScience.restartKernelMessage();
+                const yes = DataScience.restartKernelMessageYes();
+                const dontAskAgain = DataScience.restartKernelMessageDontAskAgain();
+                const no = DataScience.restartKernelMessageNo();
+
+                const response = await this.applicationShell.showInformationMessage(
+                    message,
+                    { modal: true },
+                    yes,
+                    dontAskAgain,
+                    no
+                );
+                if (response === dontAskAgain) {
+                    await this.disableAskForRestart(document.uri);
+                    void this.wrapKernelMethod('restart', kernel);
+                } else if (response === yes) {
+                    void this.wrapKernelMethod('restart', kernel);
+                }
+            } else {
+                void this.wrapKernelMethod('restart', kernel);
+            }
+        }
+    }
+
+    private readonly pendingRestartInterrupt = new WeakMap<IKernel, Promise<void>>();
+    private async wrapKernelMethod(currentContext: 'interrupt' | 'restart', kernel: IKernel) {
+        const notebook = getAssociatedNotebookDocument(kernel);
+        if (!notebook) {
+            throw new Error('Unable to start a kernel that is not attached to a notebook document');
+        }
+        // We don't want to create multiple restarts/interrupt requests for the same kernel.
+        const pendingPromise = this.pendingRestartInterrupt.get(kernel);
+        if (pendingPromise) {
+            return pendingPromise;
+        }
+        const promise = (async () => {
+            // Get currently executing cell and controller
+            const currentCell = kernel.pendingCells[0];
+            const controller = this.notebookControllerManager.getSelectedNotebookController(notebook);
+            try {
+                if (!controller) {
+                    throw new Error('No kernel associated with the notebook');
+                }
+                // Wrap the restart/interrupt in a loop that allows the user to switch
+                await KernelConnector.wrapKernelMethod(
+                    controller.controller,
+                    controller.connection,
+                    currentContext,
+                    kernel.creator,
+                    this.serviceContainer,
+                    { resource: kernel.resourceUri, notebook },
+                    new DisplayOptions(false),
+                    this.disposableRegistry
+                );
+            } catch (ex) {
+                if (currentCell) {
+                    await endCellAndDisplayErrorsInCell(
+                        currentCell,
+                        kernel.controller,
+                        await this.errorHandler.getErrorMessageForDisplayInCell(ex, currentContext),
+                        false
+                    );
+                } else {
+                    void this.applicationShell.showErrorMessage(ex.toString());
+                }
+            }
+        })();
+        promise.finally(() => {
+            if (this.pendingRestartInterrupt.get(kernel) === promise) {
+                this.pendingRestartInterrupt.delete(kernel);
+            }
+        });
+        this.pendingRestartInterrupt.set(kernel, promise);
+        return promise;
+    }
+
+    private async shouldAskForRestart(notebookUri: Uri): Promise<boolean> {
+        if (this.kernelInterruptedDontAskToRestart) {
+            return false;
+        }
+        const settings = this.configurationService.getSettings(notebookUri);
+        return settings && settings.askForKernelRestart === true;
+    }
+
+    private async disableAskForRestart(notebookUri: Uri): Promise<void> {
+        const settings = this.configurationService.getSettings(notebookUri);
+        if (settings) {
+            this.configurationService
+                .updateSetting('askForKernelRestart', false, undefined, ConfigurationTarget.Global)
+                .ignoreErrors();
+        }
     }
 }
